fix: Router 무한 루프 개선
