--- src/Root.tsx
@@ -1,14 +1,0 @@
-import GlobalStyles from "@/components/GlobalStyles";
-
-import App from "./App";
-
-function Root() {
-	return (
-		<>
-			<GlobalStyles />
-			<App />
-		</>
-	);
-}
-
-export default Root;
