+++ src/init/BaseLayoutWithChildren.tsx
@@ -1,0 +1,14 @@
+import Header from "@/components/Header";
+
+interface Props {}
+
+const BaseLayout = ({ children }: React.PropsWithChildren<Props>) => {
+	return (
+		<>
+			<Header />
+			{children}
+		</>
+	);
+};
+
+export default BaseLayout;
